1. Removed most of the doxygen style comments because I felt that they just made the headers read more difficult, and didn't have much value anyways.
2. Rewrote most of the printing methods back to overloading the <<
operator. That feels more natural after all, and one can always print
into a string stream and convert it into a string if necessary.

3. Renamed and redesigned UnitIntervalPoint (now Mod1Number) class.
Previously the < operator wasn't transitive because points that where
too close were considered equal. It was confusing. Whether too points
are too close will be checked elsewhere.

4. Started designing the Mod1NumberIntExchange class which keeps track
of how a number can be written as an integer linear combination of the
lengths in an interval exchange.