1. The whole project is added to the balazs namespace.
2. Finished refactoring the intex folder. Added a LengthsAndTwist class
which serve as an intermediate class between TwistedIntervalExchangeMap
and Mod1NumberIntExchange. This way the other two class don't depend on
each other as much.

3. Got rid of the IntervalExchangeBase class and only left the
TwistedIntervalExchangeMap class. The IntervalExchangeFoliationDisk is
renamed to IntervalPairing and implemented using in terms of
TwistedIntervalExchangeMap (using containment).

4. The WeighedTree class is also polished. Threw out the
m_numDescendants member of WeighedTree::Node, because it wasn't really
needed.

5. In some other classes (e.g. Modint), some member or friend functions
are removed and transformed into non-member non-friends instead.

Next the foliation classes need to be cleaned up.